requirements.txt
fpdf==1.7.2
gradio==3.33.1
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
openai==0.27.6
python-dotenv==1.0.0